vk-part-master/requirements.txt
python-dotenv~=0.15.0
pika~=1.1.0
requests~=2.24.0
vk-api~=11.9.0